+++ app/controllers/application_controller.rb
@@ -3,7 +3,8 @@ class ApplicationController < ActionController::Base
   # For APIs, you may want to use :null_session instead.
   protect_from_forgery with: :exception
 
-  layout "contour/layouts/application"
+  # layout "contour/layouts/application"
+  layout "layouts/application_custom"
 
   include DateAndTimeParser