+++ config/initializers/themes.rb
@@ -1,0 +1,1 @@
+Slice::THEMES = [['Night (Experimental)', 'night']]